download_packages

Uses resource_map and dataone::getPackage to download the data into a BagItFile. Then utils::unzip unzips the data and stores in the user's directory.

download_package_rm(
  resource_map,
  directory,
  CNode = "PROD",
  download_format = "application/bagit-097",
  overwrite_directory = TRUE
)

Arguments

resource_map

the resource map that corresponds to the given data package

directory

location that download.packages places the data

CNode

defaults to "PROD"

download_format

typically "application/bagit-097". Other possible formats currently unknown.

overwrite_directory

boolean that indicates whether or not the function should overwrite the directory

Value

results of download