Fertilizer Nutrient Composition Table

A dataset of fertilizer and organic matter addition types and their nitrogen and carbon composition, based on the SWAT model's `fertilizer.frt` table and DayCent model defaults for organic matter C:N ratio parameters.

Usage

fertilizer_composition_data

Format

A tibble with one row per fertilizer type and the following columns:

name

character. Short identifier from SWAT (e.g., "urea", "manure").

description

character. Longer description of the fertilizer or manure type.

fraction_mineral_n

numeric. Fraction of total nitrogen in mineral form.

fraction_nh3_n

numeric. Fraction of fertilizer by mass that is ammonium-n (NH\(_3\)-N).

fraction_no3_n

numeric. Fraction of fertilizer by mass that is nitrate-N (NO\(_3\)-N). Computed as fraction_mineral_n - fraction_nh3_n.

fraction_organic_n

numeric. Fraction of organic matter that is nitrogen.

fraction_c

numeric. Fraction of mass that is carbon.

cn_ratio

numeric. Carbon-to-nitrogen ratio for organic matter. Assigned based on DayCent organic matter parameterterizations.

Source

https://github.com/swat-model/swatplus

DayCent model default parameter file: `omad.100` obtained from the Soil Carbon Solutions Center, https://www.soilcarbonsolutionscenter.com

Details

This table is based on SWAT model's fertilizer.frt file, and uses C:N ratios (cn_ratio) from DayCent model default parameter files. fraction_nh3_n and fraction_no3_n represent the fraction of fertilizer by mass that is ammonium-N and nitrate-N, respectively. This is different from the SWAT model's definition of fraction_nh3_n as a fraction of the total mineral N.