Load/extract + match species module

Load/extract + match species module

get_veg_module(
  input_veg,
  outfolder,
  start_date,
  end_date,
  dbparms,
  new_site,
  host,
  machine_host,
  overwrite
)

Arguments

input_veg

list, this is a sublist of settings$run$inputs that has info about source, id, metadata of the requested IC file

outfolder

path to where the processed files will be written

start_date

date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB

end_date

date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB

dbparms

list, settings$database info reqired for opening a connection to DB

new_site

data frame, id/lat/lon/name info about the site

host

list, host info as in settings$host, host$name forced to be "localhost" upstream

machine_host

local machine hostname, e.g. "pecan2.bu.edu"

overwrite

logical flag for convert_input

Author

Istem Fer