Match species to PFTs + veg2model module

Match species to PFTs + veg2model module

put_veg_module(getveg.id, dbparms, input_veg, pfts, outfolder, n.ensemble,
  dir, machine, model, start_date, end_date, new_site, host, overwrite)

Arguments

getveg.id
list, input.id and dbfile.id of the IC file in intermediate pecan standard

dbparms
list, settings$database info reqired for opening a connection to DB

input_veg
list, this is a sublist of settings$run$inputs that has info about source, id, metadata of the requested IC file

pfts
list, same as settings$pfts

outfolder
path to where the processed files will be written

n.ensemble
integer, ensemble member number

dir
dir path to dbfiles on local machine

machine
data frame, DB info regarding localhost machine id/hostname etc.

model
model name, e.g. "ED2"

start_date
date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$start.date, otherwise start_date of the IC file in DB

end_date
date in "YYYY-MM-DD" format, in case of source==FIA it's the settings$run$end.date, otherwise end_date of the IC file in DB

new_site
data frame, id/lat/lon/name info about the site

host
list, host info as in settings$host, host$name forced to be "localhost" upstream

overwrite
logical flag for convert.input