sample_ic

sample_ic

Usage

sample_ic(
  in.path,
  in.name,
  start_date,
  end_date,
  outfolder,
  n.ensemble,
  machine_host,
  source,
  bin_var = "DBH",
  bin_size = 10,
  bin_herb_soil = TRUE,
  ...
)

Arguments

in.path

path to folder of the file to be sampled

in.name

file name of the file to be sampled

start_date

date in "YYYY-MM-DD" format

end_date

date in "YYYY-MM-DD" format

outfolder

dir path, whete to write the file

n.ensemble

integer, ensemble member number

machine_host

localhost name, e.g. "pecan2.bu.edu"

source

string to appear in file names, e.g. "PalEON"

bin_var

variable you would like to sample by, DEFAULT is DBH

bin_size

bin size for sampling, DEFAULT is 10

bin_herb_soil

if we want to use bin size for both herb and soil sampling

...

Other inputs

Author

Istem Fer