Estimate soil parameters from texture class or sand/silt/clay

Estimate soil parameters from texture class or sand/silt/clay

soil_params(
  soil_type = NULL,
  sand = NULL,
  silt = NULL,
  clay = NULL,
  bulk = NULL
)

Arguments

soil_type

USDA Soil Class. See Details

sand

percent sand

silt

percent silt

clay

percent clay

bulk

soil bulk density (optional, kg m-3)

Value

list of soil hydraulic and thermal parameters

Details

* Specify _either_ soil_type or sand/silt/clay. soil_type will be ignored if sand/silt/clay is provided * If only 2 out of sand/silt/clay are provided, it will be assumed they sum to 100 * Valid soil class options: "Sand","Loamy sand","Sandy loam","Silt loam","Loam", "Sandy clay loam","Silty clay loam","Clayey loam", "Sandy clay","Silty clay","Clay","Peat","Bedrock", "Silt","Heavy clay","Clayey sand","Clayey silt" * Based on ED2/R-utils/soilutils.r * Hydraulics based on Cosby et al 1984, using table 4 and equation 1 (which is incorrect it should be saturated moisture potential over moisture potential)

Examples

sand <- c(0.3, 0.4, 0.5)
clay <- c(0.3, 0.3, 0.3)
soil_params(sand=sand,clay=clay)
#> $soil_n
#> [1] 8 8 6
#> 
#> $soil_type
#> [1] "Clayey loam"     "Clayey loam"     "Sandy clay loam"
#> 
#> $fraction_of_sand_in_soil
#> [1] 0.3 0.4 0.5
#> 
#> $fraction_of_clay_in_soil
#> [1] 0.3 0.3 0.3
#> 
#> $fraction_of_silt_in_soil
#> [1] 0.4 0.3 0.2
#> 
#> $soil_hydraulic_b
#> [1] 7.72 7.69 7.66
#> 
#> $soil_water_potential_at_saturation
#> [1] -0.3213661 -0.2233572 -0.1552387
#> 
#> $soil_hydraulic_conductivity_at_saturation
#> [1] 2.719764e-06 3.635224e-06 4.858825e-06
#> 
#> $volume_fraction_of_water_in_soil_at_saturation
#> [1] 0.4513 0.4371 0.4229
#> 
#> $volume_fraction_of_water_in_soil_at_field_capacity
#> [1] 0.2962469 0.2820442 0.2682089
#> 
#> $volume_fraction_of_condensed_water_in_soil_at_wilting_point
#> [1] 0.2030625 0.1870019 0.1719619
#> 
#> $slpotwp
#> [1] -152.9574 -152.9574 -152.9574
#> 
#> $volume_fraction_of_condensed_water_in_dry_soil
#> [1] 0.1848382 0.1701565 0.1564135
#> 
#> $slpotcp
#> [1] -316.112 -316.112 -316.112
#> 
#> $slcpd
#> [1] 1238303 1263112 1287556
#> 
#> $thcond0
#> [1] 0.7823956 0.8063304 0.8304992
#> 
#> $thcond1
#> [1] 0.5333047 0.5333047 0.5333047
#> 
#> $thcond2
#> [1] 0.8136880 0.7926926 0.7714920
#> 
#> $thcond3
#> [1] -0.4678112 -0.4678112 -0.4678112
#> 
#> $soil_thermal_conductivity
#> [1] 1.211425 1.258009 1.308734
#> 
#> $soil_thermal_conductivity_at_saturation
#> [1] 1.697869 1.767113 1.840888
#> 
#> $soil_albedo
#> [1] 0.25 0.25 0.25
#> 
#> $soil_bulk_density
#> [1] 1350 1350 1350
#> 
#> $slden
#> [1] 1600 1600 1600
#> 
#> $soil_thermal_capacity
#> [1] 917.2615 935.6388 953.7453
#>