A function to estimate the soil parameters based on SoilGrids soil texture data and write the parameter paths into settings

A function to estimate the soil parameters based on SoilGrids soil texture data and write the parameter paths into settings

Usage

soil_params_ensemble_soilgrids(
  settings,
  sand,
  clay,
  silt,
  outdir,
  write_into_settings = TRUE
)

Arguments

settings

A multi-site settings

sand, clay, silt

Data frames containing fraction in percentage from SoilGrids250m v2.0, each with columns "Depth", "Quantile", "Siteid", and "Value"

outdir

Provide the path to store the parameter files

write_into_settings

Whether to write the path of parameter file into the setting. The default is TRUE

Value

Ensemble soil parameter files defined in outdir and file paths in xml file

Author

Qianyu Li

Examples

if (FALSE) { # \dontrun{

outdir <- "/projectnb/dietzelab/Cherry/SoilGrids_texture/39NEON"
# each file contains percent salt, silt, or clay
sand <- readRDS("/path/to/SoilGrids_texture/sand_percent.rds")
clay <- readRDS("/path/to/SoilGrids_texture/clay_percent.rds")
silt <- readRDS("/path/to/SoilGrids_texture/silt_percent.rds")
settings <-read.settings("/path/to/pecan_monthly_SDA_soilwater.xml")
soil_params_ensemble_soilgrids(settings,sand,clay,silt,outdir)
} # }