Biomass and soil data from FluxNet sites

Contains data from 246 Fluxnet sites. Variables include aboveground and belowground biomass in various pools, plus soil texture/chemistry/horizonation/C&N stocks.

Usage

BADM

Format

## `BADM` A data frame with 12,300 rows and 13 columns:

SITE_ID

Fluxnet code for the site

LOCATION_ELEV, LOCATION_LAT, LOCATION_LON

site coordinates

Date

Measurement date

GROUP_ID

TODO

VARIABLE_GROUP

category, eg abovground biomass or soil chemistry

VARIABLE, DATAVALUE

key and value for each measured variable

NA_L1CODE, NA_L1NAME, NA_L2CODE, NA_L2NAME

numeric IDs and names for the Level 1 and level 2 ecoregions where this site is located

Source

Originally from Fluxnet <https://fluxnet.org/badm-data-product/>, but the provenence and age of this specific file is not clear.