InventoryGrowthFusion

this code fuses forest inventory data with tree growth data (tree ring or dendrometer band) for the same plots. Code is a rewrite of Clark et al 2007 Ecol Appl into JAGS

Usage

InventoryGrowthFusion(
  data,
  cov.data = NULL,
  time_data = NULL,
  n.iter = 5000,
  n.chunk = n.iter,
  n.burn = min(n.chunk, 2000),
  random = NULL,
  fixed = NULL,
  time_varying = NULL,
  burnin_plot = FALSE,
  save.jags = "IGF.txt",
  z0 = NULL,
  save.state = TRUE,
  restart = NULL
)

Arguments

data

list of data inputs

cov.data

covariate data

time_data

required if time_varying is provided

n.iter

total number of iterations across all chunks

n.chunk

number of MCMC steps to evaluate at a time. Will only return LAST. If restarting, second number in vector is chunk to start from

n.burn

number of steps to automatically discard as burn-in

random

whether or not to include random effects

fixed

formula for fixed effects

time_varying

formula for time-varying effects

burnin_plot

logical: display a plot of the burnin steps?

save.jags

logical: Save the generated JAGS script?

z0

initial conditions for state variable

save.state

whether or not to include inferred DBH in output (can be large). Enter numeric value to save.state periodically (in terms of n.chunk)

restart

final mcmc.list from previous execution. NULL for new run. TRUE to save final state for new run.

Value

an mcmc.list object

Note

Requires JAGS