Download NEON Soil Water Content and Soil Salinity data by date and site name

Download NEON Soil Water Content and Soil Salinity data by date and site name

Usage

download_NEON_soilmoist(
  site,
  avg = "all",
  var = "all",
  startdate = NA,
  enddate = NA,
  outdir
)

Arguments

site

four letter NEON site code name(s). If no site is specified, it will download all of them (chr) (e.g "BART" or c("SRER", "KONA", "BART"))

avg

averaging interval (minutes): 1, 30, or both ("all") . default returns both

var

variable of interest: "SWC" (soil water content) or "SIC" (soil ion content) or both ("all") default returns both. Both variables will be saved in outdir automatically (chr)

startdate

start date as YYYY-mm. If left empty, all data available will be downloaded (chr)

enddate

start date as YYYY-mm. If left empty, all data available will be downloaded (chr)

outdir

out directory to store the following data: .rds list files of SWC and SIC data for each site and sensor position, sensor positions .csv for each site, variable description .csv file, readme .csv file

Value

List of specified variable(s) AND prints the path to output folder

Author

Juliette Bateman

Examples

if (FALSE) { # \dontrun{
test <- download_NEON_soilmoisture(
  site = c("SRER", "BART", "KONA"),
  avg = 30,
  var = "SWC",
  startdate = "2019-01",
  enddate = "2020-01",
  outdir = getwd())} # }