Default parameters for calculating soil properties from sand & clay content

Default parameters for calculating soil properties from sand & clay content

Usage

soil_class

Format

## `soil_class` A list with 26 entries:

air.cond, h2o.cond, sand.cond, silt.cond, clay.cond

thermal conductivity, W m^-1 K^-1

air.hcap, sand.hcap, silt.hcap, clay.hcap

heat capacity, J m^-3 K^-1

kair, ksand, ksilt, kclay

relative conductivity factor

fieldcp.K

hydraulic conductance at field capacity, mm day^-1

grav

gravity acceleration, m s^-2

soil.key

Abbreviations for each of 18 soil texture classes, e.g. "SiL", "LSa"

soil.name

Names for 18 soil texture classes, e.g. "Sand", "Silty clay"

soilcp.MPa

soil water potential when air-dry, MPa

soilld.MPa

soil water potential at critical water content, MPa

soilwp.MPa

soil water potential at wilting point, MPa

stext.lines

list of 18 lists, each giving minimum and maximum sand/silt/clay contents for a soil texture class

stext.polygon

list of 18 lists, each giving corner points in the soil texture triangle for a soil texture class

texture

data frame with 13 rows and 21 columns, giving default parameter values for 13 named soil textures

theta.crit

critical water content (fractional soil moisture at which plants start dropping leaves), m^3 m^-3

xclay.def

default volume fraction of sand in each of 18 soil texture classes

xsand.def

default volume fraction of clay in each of 18 soil texture classes

Source

The hydraulic parameters are derived from Cosby et al 1984, "A Statistical Exploration of the Relationships of Soil Moisture Characteristics to the Physical Properties of Soils", Water Resources Research 20(6): 682-690. This implementation comes from one provided by the ED2 model, plus `texture.csv` from a source not recorded. Package `PEcAn.linkages` contains an identical texture.csv, also with no obvious source label. See also comments in soil_utils.R